bruk kun autogenererte dtos
