+++ packages/types/src/Ytelse.ts
@@ -1,0 +1,1 @@
+export type Ytelse = 'FORELDREPENGER' | 'SVANGERSKAPSPENGER' | 'ENGANGSSTØNAD';
